gradle: move mapping-list codec to its own dependency-free subpackage
Move EncodeMappingList/DecodeMappingList out of the gradle package into
gradle/mappinglist (as Encode/Decode). The gradle package pulls in go-utils and
go-glob for running Gradle; a consumer that only needs the mapping-list codec
(e.g. the Google Play Deploy step, which never runs Gradle) should not have to
depend on all of that. The new subpackage has no dependencies beyond the
standard library, so module graph pruning keeps it cheap to import.

Variant parsing stays in the gradle package, where it composes with Artifact.
